Move settings to their own Vuex module
