+++ frontend/model/settings/colors.js
