+++ frontend/views/utils/avatar.js
@@ -2,7 +2,7 @@
 
 import sbp from '@sbp/sbp'
 import blockies from '@utils/blockies.js'
-import Colors from '~/frontend/model/colors.js'
+import Colors from '~/frontend/model/settings/colors.js'
 import { randomFromArray } from '@model/contracts/shared/giLodash.js'
 import { imageDataURItoBlob } from '@utils/image.js'
